aider: Create a file called config.rs in the root directory that holds configuration structs and a utility function called authorize_user() that validates the user cookie by making a call to the postgres database instance.
# Aider chat conversation:

USER: Create a file called config.rs in the root directory that will hold all of our configuration structs and some utility functions that will be regularly used by other crates. Create one function called authorize_user() that will return a AuthorizedUser struct and will validate that the user cookie is valid by making a call to the postgres database instance that is found in the main.rs file.

USER: Create a file called config.rs in the root directory that will hold all of our configuration structs and some utility functions that will be regularly used by other crates. Create one function called authorize_user() that will return a AuthorizedUser struct and will validate that the user cookie is valid by making a call to the postgres database instance that is found in the main.rs file.

USER: Create a file called config.rs in the root directory that will hold all of our configuration structs and some utility functions that will be regularly used by other crates. Create one function called authorize_user() that will return a AuthorizedUser struct and will validate that the user cookie is valid by making a call to the postgres database instance that is found in the main.rs file.

ASSISTANT: No changes are needed.